Improve the way the text is displayed (#12)
* Add functions to improve text display

* Identify minor typo

* Use rich library to output all text

Include a console.py file where the Console() is used and replanced all
print() with console.print()

* Add optional argument to change terminal theme

* Allow incorrect input for choice of move

* Improve the use of rich.console

* Update library version

* Improve terminal backgournd feature

Ensure that pokemon id is less than 719 and fix the names of some
specific pokemon to match the pokemon-terminal library

* Add rich to library dependecies

* Change black usage in tox.ini

* Make sure that all tests and checks pass

* Text is printed on top of the previous one

In order to give a different effect for the displayed text,
it is now printed on top of the previous sentence.
i.e. instead of printing statement_1 and below it statement_2,
we now print statement_1, then delete statement_1 and then
print statement_2 at its place

* Fix minor bug

* Add while loop for "choose pokemon"

When pokemon faints include a while loop to choose a new one
that makes sure that incorrect input is given. Previously when an
invalid input was given the game crashed

* Animate displayed text

* Fix minor issue with pylint

* Update version and CHANGELOG.md